# server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2",
    "title": "Plant-app-backend",
    "description": "REST API documentation"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/login": {
      "post": {
        "description": "User login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "UserName": {
                  "type": "string",
                  "example": "string"
                },
                "PasswordHash": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/register": {
      "post": {
        "description": "User registration",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "UserName": {
                  "type": "string",
                  "example": "string"
                },
                "PasswordHash": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/plants": {
      "get": {
        "description": "Fetches plant records linked to user",
        "parameters": [
          {
            "name": "UserId",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "creates plant record linked to user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "UserId": {
                  "type": "number",
                  "example": 0
                },
                "NamePlant": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "put": {
        "description": "edits plant record linked to user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "PlantId": {
                  "type": "number",
                  "example": 0
                },
                "UserId": {
                  "type": "number",
                  "example": 0
                },
                "PresetId": {
                  "type": "number",
                  "example": 0
                },
                "NamePlant": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "delete": {
        "description": "removes plant record linked to user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "UserId": {
                  "type": "number",
                  "example": 0
                },
                "NamePlant": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/preset": {
      "get": {
        "description": "fetches preset table linked to user",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "creates plant record linked to user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "NamePreset": {
                  "type": "string",
                  "example": "string"
                },
                "Temp": {
                  "type": "number",
                  "example": 0
                },
                "Moist": {
                  "type": "number",
                  "example": 0
                },
                "AirQuality": {
                  "type": "number",
                  "example": 0
                },
                "UserId": {
                  "type": "number",
                  "example": 0
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "put": {
        "description": "edits preset record linked to user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Id": {
                  "type": "number",
                  "example": 0
                },
                "NamePreset": {
                  "type": "string",
                  "example": "string"
                },
                "Temp": {
                  "type": "number",
                  "example": 0
                },
                "Moist": {
                  "type": "number",
                  "example": 0
                },
                "AirQuality": {
                  "type": "number",
                  "example": 0
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "delete": {
        "description": "removes plant record linked to user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Id": {
                  "type": "number",
                  "example": 0
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/history": {
      "get": {
        "description": "fetches sensor history linked to plant",
        "parameters": [
          {
            "name": "PlantId",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/test": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}